feat: centralized Zod validation middleware (validateOrThrow) and schema extraction
- Created lib/validations/validateOrThrow.js — throw-based validation utility
  that composes with withErrorHandler, returning 400 + structured field-level
  error details on Zod parse failure.
- Created 8 new schema files in lib/validations/ extracting inline Zod schemas
  from route handlers: activities, conversations, flashcards, productivity,
  curriculum, exceptions, register, settings.
- Added syncSchema to existing lib/validations/attendance.js.
- Updated lib/validations/index.js to re-export all schemas and validateOrThrow.
- Updated 10 route handlers to use validateOrThrow + centralized schemas
  instead of inline parseJSON + safeParse + manual error formatting:
  activities POST, conversations POST, flashcards POST, productivity POST,
  productivity/session POST, curriculum/sync POST, settings PATCH,
  attendance/sync POST, exceptions/create POST, exceptions/update PUT.
- Fixed package.json JSON syntax error (duplicate vitest + missing comma).

Closes #3040
