+++ app/api/conversations/route.js
@@ -1,28 +1,14 @@
-import { z } from "zod";
 import { connectDb } from "@/lib/mongodb";
 import { requireAuth } from "@/lib/rbac";
-import { withErrorHandler, parseJSON } from "@/lib/error-handler";
+import { withErrorHandler } from "@/lib/error-handler";
 import { checkRateLimit } from "@/lib/rateLimit";
 import { jsonSuccess, jsonError } from "@/lib/api-response";
-import { ValidationError } from "@/lib/errors";
+import { conversationSchema, validateOrThrow } from "@/lib/validations";
 
 export const dynamic = "force-dynamic";
 
 const MAX_PAGE_LIMIT = 100;
 const MAX_PAYLOAD_BYTES = 1024 * 50;
-
-const conversationSchema = z
-  .object({
-    userMessage: z
-      .string()
-      .min(1, "User message is required")
-      .max(10000, "User message too long (max 10000 chars)"),
-    botMessage: z
-      .string()
-      .min(1, "Bot message is required")
-      .max(10000, "Bot message too long (max 10000 chars)"),
-  })
-  .strict();
 
 /**
  * GET /api/conversations — retrieves paginated conversation history for the authenticated user.
@@ -72,15 +58,7 @@ export const POST = withErrorHandler(async (request) => {
     return jsonError("Too many requests. Please try again later.", 429);
   }
 
-  const body = await parseJSON(request, MAX_PAYLOAD_BYTES);
-  const parsed = conversationSchema.safeParse(body);
-  if (!parsed.success) {
-    throw new ValidationError(
-      parsed.error.issues[0]?.message || "Invalid request payload"
-    );
-  }
-
-  const { userMessage, botMessage } = parsed.data;
+  const { userMessage, botMessage } = await validateOrThrow(request, conversationSchema, MAX_PAYLOAD_BYTES);
 
   const db = await connectDb();
   const conversation = {
