+++ app/api/exceptions/update/route.js
@@ -1,39 +1,19 @@
 import { NextResponse } from "next/server";
 import { connectDb } from "@/lib/mongodb";
 import { getUserProfileByEmail } from "@/lib/firebase-admin";
-import { withErrorHandler, parseJSON } from "@/lib/error-handler";
+import { withErrorHandler } from "@/lib/error-handler";
 import { requireRole } from "@/lib/rbac";
 import {
   AppError,
-  ValidationError,
   ForbiddenError,
   NotFoundError,
 } from "@/lib/errors";
 import { checkRateLimit } from "@/lib/rateLimit";
 import { ObjectId } from "mongodb";
-import { z } from "zod";
+import { exceptionUpdateSchema, validateOrThrow } from "@/lib/validations";
 
 // Required to prevent build-time static generation errors
 export const dynamic = "force-dynamic";
-
-const exceptionUpdateSchema = z.object({
-  exceptionId: z
-    .string({
-      required_error: "exceptionId is required",
-      invalid_type_error: "exceptionId is required",
-    })
-    .trim()
-    .min(1, "exceptionId is required")
-    .refine((val) => ObjectId.isValid(val), {
-      message: "Invalid exception ID",
-    }),
-  status: z.enum(["approved", "rejected"], {
-    required_error: "Invalid status value",
-    invalid_type_error: "Invalid status value",
-    message: "Invalid status value",
-  }),
-  comments: z.string().optional(),
-});
 
 export const PUT = withErrorHandler(async (request) => {
   const { payload: decodedToken, profile } = await requireRole(request, [
@@ -47,33 +27,7 @@ export const PUT = withErrorHandler(async (request) => {
   if (!rateLimitResult.allowed) {
     throw new AppError("Too many attempts. Please try again later.", 429);
   }
-  const body = await parseJSON(request, 1024 * 10);
-
-  const validation = exceptionUpdateSchema.safeParse(body);
-  if (!validation.success) {
-    let firstError =
-      validation.error.issues?.[0]?.message || "Invalid request payload";
-    const path = validation.error.issues?.[0]?.path?.[0];
-    const code = validation.error.issues?.[0]?.code;
-
-    if (
-      path === "exceptionId" &&
-      (code === "invalid_type" || firstError.includes("Required"))
-    ) {
-      firstError = "exceptionId is required";
-    } else if (
-      path === "status" &&
-      (code === "invalid_type" ||
-        code === "invalid_enum_value" ||
-        firstError.includes("Required"))
-    ) {
-      firstError = "Invalid status value";
-    }
-
-    throw new ValidationError(firstError);
-  }
-
-  const { exceptionId, status, comments } = validation.data;
+  const { exceptionId, status, comments } = await validateOrThrow(request, exceptionUpdateSchema, 1024 * 10);
 
   const db = await connectDb();
