+++ lib/validations/attendance.js
@@ -31,3 +31,20 @@ export const attendanceHeatmapQuerySchema = z.object({
   userId: z.string().min(1).max(128),
   year: z.coerce.number().int().min(2020).max(2030).optional(),
 });
+
+export const syncSchema = z.object({
+  records: z
+    .array(
+      z.object({
+        id: z.number().optional(),
+        userId: z.string(),
+        studentName: z.string().optional(),
+        email: z.string().optional(),
+        confidenceScore: z.number().optional(),
+        queuedAt: z.number(),
+        date: z.string().optional(),
+      })
+    )
+    .min(1)
+    .max(100, "Too many records in a single sync batch"),
+});
