+++ lib/validations/index.js
@@ -3,5 +3,14 @@ export * from "./auth";
 export * from "./complaints";
 export * from "./notices";
 export * from "./admin";
+export * from "./activities";
+export * from "./conversations";
+export * from "./curriculum";
+export * from "./exceptions";
+export * from "./flashcards";
+export * from "./productivity";
+export * from "./register";
+export * from "./settings";
 export { validateRequest } from "./validateRequest";
 export { withValidation } from "./withValidation";
+export { validateOrThrow } from "./validateOrThrow";
