+++ package.json
@@ -118,8 +118,6 @@
     "tailwindcss": "^4",
     "tw-animate-css": "^1.3.8",
     "vitest": "^4.1.8"
-    "vite": "6.2.0",
-    "vitest": "3.2.4"
   },
   "overrides": {
     "serialize-javascript": "7.0.5",
